+++ src/components/callToAction/persistentCta/index.ts
@@ -1,0 +1,1 @@
+export { PersistentCta } from './persistentCta';